requirements.txt
helm
torch
click
numpy
transformers
ansicolors